requirements.txt
django
python-mimeparse
python-dateutil
django-tastypie

# debug
django-extensions

# testing
factory_boy
